% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alt.desc}
\alias{alt.desc}
\title{alt.desc}
\usage{
alt.desc(msdata)
}
\arguments{
\item{msdata}{a list containing the structure of the model and data, as an output from 
\code{extractStruct.<package_name>} functions}
}
\description{
create alternative description
}
