% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bootinvFIM}
\alias{bootinvFIM}
\alias{invFIM}
\alias{bootstrap}
\title{Approximation of the inverse of the Fisher Information Matrix via parametric bootstrap}
\usage{
bootinvFIM(m, B = 1000)
}
\arguments{
\item{m}{a fitted model that will be used as the basis of the parametric bootstrap (providing the initial maximum
likelihood estimate of the parameters and the modelling framework)}

\item{B}{the size of the bootstrap sample}
}
\value{
the empirical covariance matrix of the parameter estimates obtained on the bootstrap sample
}
\description{
When the FIM is not available, this function provides an approximation of the FIM based on an estimate
of the covariance matrix of the model's parameters obtained via parametric bootstrap.
}
\author{
Charlotte Baey <\email{charlotte.baey@univ-lille.fr}>
}
