% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saemix_utilities.R
\name{bootinvFIM.saemix}
\alias{bootinvFIM.saemix}
\title{Compute the inverse of the Fisher Information Matrix using parametric bootstrap}
\usage{
\method{bootinvFIM}{saemix}(m, B = 1000)
}
\arguments{
\item{m}{the model under H1}

\item{B}{the bootstrap sample size}
}
\description{
Compute the inverse of the Fisher Information Matrix using parametric bootstrap
}
