% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiBarObject.R
\docType{class}
\name{chibarObject-class}
\alias{chibarObject-class}
\alias{chiBarObject}
\alias{chibarObject}
\title{Class "chibarObject"}
\description{
An object of the \code{chibarObject} class, storing the
}
\details{
Details of the algorithm can be found in the pdf file accompanying the package.
}
\section{Slots}{

\describe{
\item{\code{V}}{a positive-definite matrix}

\item{\code{dims}}{the set of dimensions defining the cone}

\item{\code{orthant}}{logical, equals \code{TRUE} is the cone is the nonnegative orthant of R^r}
}}

\section{Objects from the Class}{

An object of the chibarObject contains the following slots:
}

