% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiBarObject.R
\docType{class}
\name{chiBarSquareObject-class}
\alias{chiBarSquareObject-class}
\alias{chiBarSquareObject}
\title{Class "chiBarSquareObject"}
\description{
An object of the \code{chiBarSquareObject} class, storing the parameters of the chi-bar-square distribution.
}
\section{Slots}{

\describe{
\item{\code{V}}{a positive-definite matrix}

\item{\code{dims}}{the set of dimensions defining the cone}

\item{\code{orthant}}{logical, equals \code{TRUE} is the cone is the nonnegative orthant of R^r}
}}

