% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibarsquare.R
\name{dfChiBarSquare}
\alias{dfChiBarSquare}
\title{Chi-bar-square degrees of freedom computation}
\usage{
dfChiBarSquare(msdata)
}
\arguments{
\item{msdata}{a list containing the structure of the model and data, as an output from 
\code{extractStruct.<package_name>} functions}
}
\value{
a list containing the vector of the degrees of freedom of the chi-bar-square and the dimensions
of the cone of the chi-bar-square distribution
}
\description{
Computation of the degrees of freedom of the chi-bar-square
}
