% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiBarObject.R
\name{dfchisqbar}
\alias{dfchisqbar}
\alias{dfchisqbar,chiBarSquareObject-method}
\title{dfchisqbar
Computes the degrees of freedom of the chi-square involved in the mixture}
\usage{
\S4method{dfchisqbar}{chiBarSquareObject}(object)
}
\arguments{
\item{object}{a chiBarSquareObject}
}
\description{
dfchisqbar
Computes the degrees of freedom of the chi-square involved in the mixture
}
