% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_utilities.R
\name{extractFIM.lme}
\alias{extractFIM.lme}
\title{Extract FIM}
\usage{
extractFIM.lme(m, struct)
}
\arguments{
\item{m}{the model to extract the FIM from}

\item{struct}{the structure of the covariance matrix (either 'full', 'diag', or 'blockdiag)}
}
\description{
Extract FIM
}
