% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFIMnlme.R
\name{extractFIMnlme}
\alias{extractFIMnlme}
\alias{FIM}
\alias{nlme}
\title{Extraction of the Fisher Information Matrix for nlme package}
\usage{
extractFIMnlme(m, struct)
}
\arguments{
\item{m}{a model fitted using nlme}

\item{struct}{a string giving the structure of the covariance matrix: either \code{diag} for a diagonal
matrix, \code{blockDiag} for a block diagonal matrix of \code{full} for a matrix with non-zero components}
}
\value{
the FIM matrix for the variance components
}
\description{
Extraction of the Fisher Information Matix for variance components fitted with nlme using Delta method
}
\details{
This function extract the FIM computed by the nlme for the transformed variance components, and uses
the Delta method to compute the FIM for the natural variance components (i.e. variances and covariances)
}
\author{
Charlotte Baey <\email{charlotte.baey@univ-lille.fr}>
}
