% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_utilities.R
\name{extractStruct.lme}
\alias{extractStruct.lme}
\title{Extract model structure}
\usage{
\method{extractStruct}{lme}(m1, m0, randm0)
}
\arguments{
\item{m1}{the fit under H1}

\item{m0}{the fit under H0}

\item{randm0}{a boolean indicating whether random effects are present in m0}
}
\description{
Extract model structure
}
