% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_utilities.R
\name{extractVarCov.lme}
\alias{extractVarCov.lme}
\title{Extract covariance matrix}
\usage{
\method{extractVarCov}{lme}(m)
}
\arguments{
\item{m}{a fit from nlme package (either linear or nonlinear)}
}
\description{
Extract covariance matrix of the random effects for a model fitted with nlme.
}
