% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{extractVarCov}
\alias{extractVarCov}
\title{Extract covariance matrix}
\usage{
extractVarCov(m)
}
\arguments{
\item{m}{a fit from lme4 package (either linear or nonlinear)}
}
\description{
Extract covariance matrix of the random effects for a model fitted with lme4.
}
