% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fim.vctest}
\alias{fim.vctest}
\title{Extract the Fisher Information Matrix}
\usage{
fim.vctest(object)
}
\arguments{
\item{object}{an object of class vctest}
}
\description{
Extract the Fisher Information Matrix
}
