% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCO.R
\name{funcCO}
\alias{funcCO}
\alias{objFunction}
\alias{gradObjFunction}
\alias{ineqCstr}
\alias{ineqCstrDiag}
\alias{eqCstr}
\title{Internal functions for constrained minimization}
\usage{
objFunction(x, cst)

gradObjFunction(x, cst)

ineqCstr(x, cst)

ineqCstrDiag(x, cst)

eqCstr(x, cst)
}
\arguments{
\item{x}{A vector}

\item{cst}{A list of constants to be passed to the optimisation function}
}
\value{
value of the objective function, its gradient, and the set of inequality and euqality constraints
}
\description{
Groups of functions used for the constrained minimization problem arising in the computation of the
likelihood ratio test statistics.
}
