% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarCovlme4.R
\name{getVarCovlme4}
\alias{getVarCovlme4}
\title{Extract covariance matrix}
\usage{
getVarCovlme4(m)
}
\arguments{
\item{m}{a fit from lme4 package (either linear or nonlinear)}
}
\description{
Extract covariance matrix of the random effects for a model fitted with lme4.
}
