% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarCovnlme.R
\name{getVarCovnlme}
\alias{getVarCovnlme}
\title{Extract covariance matrix}
\usage{
getVarCovnlme(m)
}
\arguments{
\item{m}{a fit from nlme package (either linear or nonlinear)}
}
\description{
Extract covariance matrix of the random effects for a model fitted with nlme.
}
