% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{getVarCovnlme}
\title{Helpers for packages nlme and lme4}
\usage{
getVarCovnlme(m)
}
\arguments{
\item{m}{a fit from nlme package (either linear or nonlinear)}
}
\description{
Functions extracting features from nlme or lme4 objects.
}
