% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null.desc}
\alias{null.desc}
\title{null.desc}
\usage{
null.desc(msdata)
}
\arguments{
\item{msdata}{a list containing the structure of the model and data, as an output from 
\code{extractStruct.<package_name>} functions}
}
\description{
create null.value description
}
\details{
Useful intern functions
}
