% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgName.R
\name{pckName}
\alias{pckName}
\title{Extract package name from a fitted mixed-effects model}
\usage{
pckName(m)
}
\arguments{
\item{m}{a model with random effects fitted with \code{nlme}, \code{lme4} or \code{saemix}}
}
\value{
a string giving the name of the package
}
\description{
Extract package name from a fitted mixed-effects model
}
