% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.varTestObject.R
\name{plot.varTestObject}
\alias{plot.varTestObject}
\title{Diagnostic plot for the approximation of the chi-bar-square distribution}
\arguments{
\item{x}{a object of class \code{\link{varTestObject}} obtained from a call to function \code{\link{varTest}}}
}
\description{
Plot the empirical cumulative distribution function (cdf) of the simulated chi-bar-square distributed variable, along with the exact cdf
of all the chi-square distributions involved in the mixture, and with the cdf based on the approximated weights. This function can only be used when
the weights were approximated by simulation.
}
