% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.desc.message}
\alias{print.desc.message}
\title{print.desc.message}
\usage{
\method{print}{desc.message}(msdata)
}
\arguments{
\item{msdata}{a list containing the structure of the model and data, as an output from 
\code{extractStruct.<package_name>} functions}
}
\description{
print a message to indicate the null and alternative hypotheses
}
