% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.res.message}
\alias{print.res.message}
\title{print.res.message}
\usage{
\method{print}{res.message}(results)
}
\arguments{
\item{results}{an object of class vctest}
}
\description{
print a message with the results
}
