% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.varTestObject.R
\name{print.varTestObject}
\alias{print.varTestObject}
\title{Print basic information about the variance components test}
\arguments{
\item{x}{a object of class \code{\link{varTestObject}} obtained from a call to function \code{\link{varTest}}}
}
\description{
Displays the likelihood ratio test statistics and the p-value of the test
}
