% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.varTestObject.R
\name{summary.varTestObject}
\alias{summary.varTestObject}
\title{Summary information for the variance components test}
\arguments{
\item{x}{a object of class \code{\link{varTestObject}} obtained from a call to function \code{\link{varTest}}}
}
\value{
a list containing the following elements:
\item{\code{lrt}}{the likelihood ratio test statistics}
\item{\code{df}}{the degrees of freedom of the chi-bar distributions involved in the chi-bar-square distribution}
\item{\code{weights}}{the weights of the limiting chi-bar-square distribution}
\item{\code{pvalWeights}}{the p-value of the test calculated using the cdf of the chi-bar-square based on (approximated) weights}
\item{\code{pvalMC}}{the Monte-Carlo estimate of the p-value of the test based on the simulated chi-bar-square distribution}
}
\description{
Displays the likelihood ratio test statistics, the limiting distribution and the p-value of the test
}
