% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.vctest}
\alias{summary.vctest}
\title{Summary}
\usage{
\method{summary}{vctest}(object, ...)
}
\arguments{
\item{object}{an object of class vctest}

\item{...}{additional arguments}
}
\description{
Summary
}
