% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTest.R
\name{varTest}
\alias{varTest}
\alias{test}
\alias{lrt}
\alias{likelihood}
\title{Variance component testing}
\usage{
varTest(m1, m0, control = list(N = 5000, parallel = F, nbcores = 1, B =
  1000), pval.comp = "bounds", fim = "extract")
}
\arguments{
\item{m1}{a fit of the model under H1, obtained from \code{\link[nlme]{nlme}}, \code{\link[lme4]{lme4}}
or \code{\link[saemix]{saemix}}}

\item{m0}{a fit of the model under H0, obtained from the same package as \code{m0}}

\item{control}{(optional) a list of control options for the computation of the chi-bar-weights}
}
\value{
A list with the following components:
\item{\code{lrt}}{the likelihood ratio test statistics}
\item{\code{ddl}}{the degrees of freedom of the chi-bar distributions involved in the chi-bar-square distribution}
\item{\code{weights}}{the weights of the limiting chi-bar-square distribution}
\item{\code{pval}}{the p-value of the test}
}
\description{
Perform a likelihood ratio test to test whether a subset of the variances of the random effects
are equal to zero. The test is defined by two hypotheses, H0 and H1, and the model under H0 is
assumed to be nested within the model under H1.
}
\details{
It is possible to tests if any subset of the variances are equal to zero. However, the function does not
currently support nested random effects, and assumes that the random effects are Gaussian.
}
\references{
Baey C, Cournède P-H, Kuhn E, 2019. Asymptotic distribution of likelihood ratio test
statistics for variance components in nonlinear mixed effects models. \emph{Computational
Statistics and Data Analysis} 135:107-122.

Silvapulle  MJ, Sen PK, 2011. Constrained statistical inference: order, inequality and shape constraints.
}
\author{
Charlotte Baey <\email{charlotte.baey@univ-lille.fr}>
}
