% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTestnlme-package.R
\docType{package}
\name{varTestnlme-package}
\alias{varTestnlme-package}
\alias{varTestnlme}
\title{Variance components testing in linear and nonlinear mixed effects models}
\description{
\code{varTestnlme} implements likelihood ratio tests for variance components in general (linear and nonlinear) mixed effects models,
assuming a multivariate Gaussian distribution for the random effects.
}
\references{
Charlotte Baey, Paul-Henry Cournède, Estelle Kuhn (2019). Asymptotic distribution of likelihood ratio test statistics for
variance components in nonlinear mixed effects models. \emph{Computational Statistics and Data Analysis.}
}
\author{
Charlotte Baey (\email{charlotte.baey@univ-lille.fr})
}
\keyword{package}
