% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varTestObject.R
\docType{class}
\name{varTestObject-class}
\alias{varTestObject-class}
\alias{varTestObject}
\title{Class "varTestObject"}
\description{
An object of the \code{varTestObject} class, storing the results of the LRT
}
\section{Slots}{

\describe{
\item{\code{lrt}}{the likelihood ratio test statistics}

\item{\code{df}}{the degrees of freedom of the chi-square distributions involved in the mixture}

\item{\code{weights}}{the weights associated to the chi-square distributions involved in the mixture}

\item{\code{pvalue}}{the p-value of the LRT}
}}

\section{Objects from the Class}{

An object of the varTestObject contains the following slots:
}

